Playing around with pygame.mixer.Sound
Finally got pygame.mixer.Sound to play sound effects. Apparently, you
have to initialize pygame and open a window before you can play sounds.

Also, I started playing with the EggBot. The first print showed that
text converts to paths as outlines and not centerlines. But there is an
extension to EggBot that works around this called Hershey Text.