Scope update, music, and flashing displays
Updates to the READ ME file to better define the scope of the project.

I also found some more music files that I like. The 'That's correct' bit
has a good radio chatter part.

I also included the Adafruit_PureIO library. Somewhere between the last
test of the resistor puzzle and the current test the SevenSegment
library broke and the PureIO library fixes it.

I also added flashing of the displays for the resistor puzzle when the
solution is found. The puzzle also breaks out of the loop when the
solution is found.